Moved types.go file into the echovault package and deleted EchoVault interface as it's no longer in use
